+++ CHANGELOG.md
@@ -26,6 +26,10 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/infrastructure/branches/compare/master%0Drelease-3.0.0
 
+### Added
+- Add support for jira-actions:3
+- Add support for virtual-users:3
+
 ## [3.0.0] - 2018-11-20
 [3.0.0]: https://bitbucket.org/atlassian/infrastructure/branches/compare/release-3.0.0%0Drelease-2.4.1